# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "牧民健康应用API",
    "description": "牧民健康应用的后端API接口",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API服务器"
    }
  ],
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "timestamp": {
            "type": "integer",
            "description": "时间戳（毫秒）"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["用户认证"],
        "summary": "用户登录",
        "description": "用户使用账号密码登录系统",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["account", "password"],
                "properties": {
                  "account": {
                    "type": "string",
                    "description": "用户账号"
                  },
                  "password": {
                    "type": "string",
                    "description": "用户密码"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "token": {
                              "type": "string",
                              "description": "JWT令牌"
                            },
                            "userId": {
                              "type": "string",
                              "description": "用户ID"
                            },
                            "nickname": {
                              "type": "string",
                              "description": "用户昵称"
                            },
                            "avatar": {
                              "type": "string",
                              "description": "用户头像URL"
                            },
                            "expiresIn": {
                              "type": "integer",
                              "description": "令牌过期时间（秒）"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "账号或密码错误"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["用户认证"],
        "summary": "用户注册",
        "description": "新用户注册系统账号",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["account", "password", "confirmPassword"],
                "properties": {
                  "account": {
                    "type": "string",
                    "description": "用户账号"
                  },
                  "password": {
                    "type": "string",
                    "description": "用户密码"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "description": "确认密码"
                  },
                  "nickname": {
                    "type": "string",
                    "description": "用户昵称"
                  },
                  "phone": {
                    "type": "string",
                    "description": "手机号"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "userId": {
                              "type": "string",
                              "description": "用户ID"
                            },
                            "account": {
                              "type": "string",
                              "description": "用户账号"
                            },
                            "nickname": {
                              "type": "string",
                              "description": "用户昵称"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数错误或账号已存在"
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["用户认证"],
        "summary": "重置密码",
        "description": "用户通过手机验证码重置密码",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phone", "verifyCode", "newPassword"],
                "properties": {
                  "phone": {
                    "type": "string",
                    "description": "手机号"
                  },
                  "verifyCode": {
                    "type": "string",
                    "description": "验证码"
                  },
                  "newPassword": {
                    "type": "string",
                    "description": "新密码"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码重置成功"
          },
          "400": {
            "description": "参数错误或手机号未注册"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["用户认证"],
        "summary": "退出登录",
        "description": "用户退出登录",
        "responses": {
          "200": {
            "description": "退出成功"
          },
          "401": {
            "description": "认证失败"
          }
        }
      }
    },
...
  }
} 